requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pymongo>=4.10.0
motor>=3.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

